issue solved

+++ src/main/java/unicam/repository/ConsegnaRepository.java
@@ -6,5 +6,5 @@ import unicam.model.consegna.Consegna;
 public interface ConsegnaRepository extends JpaRepository<Consegna, Long> {
     Consegna save(Consegna consegna);
     boolean contains(Consegna consegna);
-    void remove(int idConsegna);
+    void remove(Long idConsegna);
 }
